services/monitoring/velib_exporter/requirements_velib_exporter.txt
prometheus-client
flask
mlflow-skinny
numpy
pandas